% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.data.R
\docType{data}
\name{example.data}
\alias{example.data}
\title{example.data.}
\format{
A `data.frame` with 100 rows and 2 variables:
\describe{
  \item{id}{index, 1 to 100}
  \item{ans}{sampled letter}
}
}
\source{
\url{http://www.i.made.this.com}
}
\usage{
example.data
}
\description{
An examlpe dataset, nothing particularly useful.
}
\keyword{datasets}
