% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFlightDistance.R
\name{getFlightDistance}
\alias{getFlightDistance}
\title{getFlightDistance}
\usage{
getFlightDistance(ref1, ref2, units = "km", verbose = TRUE)
}
\arguments{
\item{ref1, ref2}{IATA code for department and arrival airports, respectively (character).}

\item{units}{units to report distance in, see below.}

\item{verbose}{if TRUE, function returns a report table else just answer in
requested units}
}
\value{
If ref1 and ref2 are matched to known airports, the function outputs
reported distance
}
\description{
getFlightDistance extracts aircraft flight distance from an online source, typically a API...
}
\note{
This version of the function extracts results from the webflyer.com
mileage calculator: http://www.webflyer.com/travel/mileage_calculator
}
\examples{
#flight distance Los Angeles to London Heathrow
getFlightDistance("LAX", "LHR")
}
\keyword{methods}
