% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/just.testing.R
\docType{package}
\name{just.testing}
\alias{just.testing}
\alias{just.testing-package}
\title{just.testing: Created to Test R Code Packaging Methods}
\description{
The just.testing package was created to test and 
demonstrate different options when building R packages. It 
is meant to be used only as an example package.
}
\section{just.testing functions}{

The foo functions: footnote.cite, getFlightDistance
}

